py2/requirements.txt
python-decouple
telethon
crypty
py-mon
watchdog
requests
pyshorteners